backend_2/requirements.txt
flask
flask-cors
pandas
spacy
google-generativeai
sentence-transformers
scikit-learn
numpy